Move time_to_ms functions to capi.rs.
These are in theory generally useful, but for the moment
they're private and only called from the capi, so it's more
clean to have them there.
